daxctl: Opt-in to /sys/bus/dax ABI
In support of the kernel's conversion of the dax-subsystem from a
'class' to a 'bus', teach the libdaxctl subsystem-layout-specific code
to parse the new layout. The kernel changes do not effect the primary
ndctl use case of putting namespaces into 'dax' mode since that uses
libnvdimm namespace device relative paths, but it does break 'ndctl list
-X' and 'daxctl list'. For that reason the kernel provides a
dax_pmem_compat driver to support the old layout and give time for
userspace components to switch.

Installation of the latest libdaxctl package arranges for a daxctl
configuration file to be dropped in /etc/modprobe.d. The modprobe
configuration blacklists dax_pmem_compat modules and sets the proper
alias for the dax_pmem module. The modprobe configuration upgrades the
default kernel handling to the /sys/bus/dax scheme.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>